UI: disable browser context menu by default on stack viewer
So far a right mouse button click on a stack viewer window triggered the
display of a browser context menu. It is however not really useful on
CATMAID's stack viewer, which is why it is not disabled by default.
Instead, the right mouse button can now be used for panning as well.

The browser context menu is still shown on other widgets that consist of
regular DOM elements.