Tile Source: Same-origin canary tile checks
Previously, credentials/headers/cookies would not be used when the
canary tile request was sent to the same origin (e.g. an HDF5 file
hosted directly by the CATMAID instance). This would cause CATMAID to
believe that it was not CORS accessible and so switch to a DOM tile
layer. Now it doesn't.