Custom mirror dialog: add text and link to add SSL exception
In case a self-signed certificate is used for the custom mirror, a SSL
exception has to be added. If such a mirror is added without having
tried to access a tile from it in a separate browser window, the mirror
can't be used because requested tiles are dropped due to the untrusted
certificate. To circumvent this, a link is added to the dialog which,
when clicked, opens a new browser window with the canary tile of the
custom mirror. This can of course be used to test if the mirror is
reachable in the first place, but given it is, a SSL exception can be
added through the browser. After this is done, the mirror can be added
and tiles will be shown.