Statistics widget: fix history table API
This statistics back-end calculates the difference in days between start
and end date. If the start date has a different UTC offset than the
end-date, this difference does not match full days anymore. This can
happen for instance if the start date is in EST time and the end date is
in EDT time). To counter this, the day differences is incremented by one
day if the exact time difference is different from the day-only
difference.

Fixes catmaid/CATMAID#1425