Tracing tool: support tagging with a custom tag set using Alt+T
Users can now configure a personal set of tags through the Settings
Widget using the "Personal tag set" field in the Tracing section. This
set of tags is persisted using the DataStore infrastructure. If Alt+T
is used on an active node, the node is tagged with this personal tag
set.