Add support for widget window aliases
A new title bar icon is added that opens a general window settings panel
similar to the available subscription settings. The displayed panel
allows to give an alias to a window, which is shown in parentheses after
the regular window title. For the sake of simplicity, this alias is not
stored and the title is modified directly.

Closes catmaid/CATMAID#1519