Set default Z for new stack viewers to first non-broken section
Due to the way screen redrawing for a new stack viewer works in CATMAID,
a stack viewer is drawn before it was moved for the first time. This
move makes sure broken sections are respected and ignored, but since
redraws happen already before, the default location is rendered as well.

This is a problem if the first section is marked broken, because tiles
for the first section are requested regardless due the default section
of a new stack viewer is zero. This tile request triggers 404 errors and
potentially CORS errors. Thanks to @davidhildebrand for the pointer to
this problem.

This is fixed by setting a new stack viewer's default Z location to the
first non-broken section. This change is consistent with the
initialization of the stack viewer's X and Y location, which are set to
the stack's center.