Support more fine-grained review status colors
Review status colors can now be more fine-grained. The Settings Widget
contains now the option "Use detail review color status" as part of the
Tracing section. If enabled, ten different colors plus the regular 0%
and 100% color will be used instead of three. This setting is optional
and the default behavior is still to use only the regular three colors.