Connector back-end: fix wrong results in /info endpoint
The connector/info endpoint updated in d6218a11153 accidentally broke
this API if not both 'pre' and 'post' parameters were provided. If for
instance only a post-skeleton was passed in, it is expected that only
those links are returned that have the respective post-skeleton as end
and any pre-skeleton as beginning. This was broken, which caused all
links to or from the result connector where returned with the passed
post-skeleton as end. This is fixed now and only the expected links
retrurned. A regression test has been added.

Fixes catmaid/CATMAID#1554