Do not print node info to the status bar upon activating a node. Instead, the status bar is cleared. To show the status info, push 'i' (just like it was possible before).
This rather minor edit removes one server request out of a total of
two for every request to update the nodes (e.g. browsing to the
next section), therefore it divides in half the e.g. transatlantic
overheads when tracing from overseas.

If the node info is at all useful when not explicitly requested,
it could be fetched together with the '/node/list' request.

Thanks to Philipp Schlegel for bringing this issue to my attention.