3D viewer: add broken/custom section node location interpolation
Nodes on specific section of the focused stack can now be automatically
moved to an interpolated location. The "Skeleton filters" tab provides
now options to either interpolate nodes on broken sections or a custom
list of sections.  The new location in the average X/Y location of the
parent and child. If only parent or child is available (leaf or parent),
its X/Y location is used.

This is useful for creating nicer renderings with nodes that are
completely off.