UI: use event key values instead of key codes
Using the keyCode property of an event is deprecated and the key field
should be used instead. The key field contains key values that are
described here:

  https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/key/Key_Values

This is more readable and needs less exceptions. While this was not a
big problem with the key code based implementation, it fixes some
problems users reported where media keys (e.g. mute) generated key codes
that CATMAID listened to. It also provides a good opportunity to add
cleaner and more robust differentiation between key combinations, e.g.
lower case and upper case letters as key shortcuts for actions. So far
actions had to explicitly exclude modifier key combinations they didn't
want to run for. This changes so that actions aren't run anymore if the
current key combination isn't explicitly listed. Given that we are
almost out of options for new keys, we should make it easier to add new
ones.

Depending on the use case, 'Shift + 3' or '#' (US layout) is more
useful. To support both (and also e.g. "Shift + A"), supported key
combination had to be made explicit. So to be able to register different
actions for e.g. "Shift + g" and "g", it is now required that all
actions list their supported actions explicitly. An action that lists
only "Shift + g" won't be called anymore if only "g" is pressed. While
this is more restrictive and makes a few actions have key lists of five
or more elements, it makes it easier to list used key combinations and
find unused key combinations. Plus it is makes calling accidentally
wrong actions for multi-use keys less likely.

Since the Bookmark Widget uses key codes to store bookmarks in the data
store, it is left untouched for now. It manages events completely
separate from CATMAID's general key event handling.