Navigator tool: allow section step size adjustment with Ctrl + #
When Ctrl + # is pressed, a dialog opens asking for a new major section
step size. Entering a number and pressing enter will set the new step
size and close the dialog. This makes it very easy to quickly jump
through a data set in larger steps with the keyboard only.

This is also called through the Tracing Tool, because the Tracing Tool
now ignores Shift + [1-5] key combinations.