Neuron navigator: replace "last used" with "annotated on" for annotations
Instead of a "Last used" column annotation tables now show an "Annotated
on" column. The last used date used before reflected the last change of
an annotation itself, which is updated during a split or merge. now the
creation time of the link relation is displayed, which doesn't change
during a merge or split operation, unlike its "edition time". The
original annotation date seems more useful to tracers when looking at
annotations on a particular neuron.

For unconstrained annotation lists, this will still show the most recent
use.