Keep annotation creation_time and edition_time during splits
So far, if skeletons have been split and annotations have been copied
over to the new skeleton (i.e. not the remaining part), annotations were
linked without any reference to the past. The link creation_time and
edition_time were therefore the time of the split operation. This
changes now so that annotation links on the new skeleton will have the
same creation_time and edition_time as they had on the skeleton before
the split.