Keep oldest annotation date when merging shared annotation
When merging two skeletons, annotations can be merged as well. If both
skeletons share a kept annotation, so far the link from the winning
skeleton has been used. This is changed now so that the oldest
annotation link is kept. The oldest annotation date seems more useful
practically when tracking who labeled a particular neuron a certain way
first.