Add simple cookie based widget state saving infrastructure
Each widget can now register for state saving, either by calling
CATMAID.registerState() directly or providing a "state" field in the
options for registerWidget(). When widgets are closed, their state is
saved by default into a cookie.  When widgets use the
getWidgetConfiguration() API, existing state is loaded when the widget
is opened. For unmanaged widgets, the loading method has to be called
manually.

The type of serialization used can be changed in a single place if
needed, currently JSON is used. It should be possible to use this
infrastructure also together with other storage types, like back-end
storage.  For now, all information is stored in browser cookies.

See catmaid/CATMADI#199