Review widget: add option to disable sending review updates to server
The checkbox "Save review updates" allows to disable sending review
updates to the server. Ending review or closing the widget will cause
all review updates to be lost. If disabled, a warning is displayed and
own revisions can't be reset. This is useful to only use the
"navigational" features of the review widget.

While these navigational features should ideally have their own widget,
this provided an easy way to provide a basic version of this
functionality. Despite this, it is also useful to have this setting as
part of the review widget, because it allows to easily demo aspects of
the review process without actually making changes.