Color picker: respect initial alpha for button background color
If a button (or other element) is set to be a color button and an
initial color is provided, its background-color is set accordingly. So
far, however, this background-color didn't respect an initial alpha
value, which made the button color look different from the actual color,
if transparency was used. This is fixed now.