Docker: only create base image which runs CATMAID with uWSGI
This image can be used both as a both image for a standalone CATMAID
image or as part of a docker-compose configuration. The CATMAID
configuration has been moved out of the Dockerfile as well and lives now
in a new entrypoint file. This provides more flexibility to work in both
scenarios.

Also the createuser.sh script has been modified to only create a new
user if they don't exist already.

See catmaid/CATMAID#1593