Graph widget: fix re-layout button
The handler for this button got recently a new callback parameter, but
due to the way the handler is bound to button, this parameter was set to
the button's mouse event. This is now fixed by explicitly binding the
callback parameter to null for the button click.

Thanks to Saba Ali for the pointer.