Review widget: allow constraining visible reviewer columns
The set of displayed columns can now be constrained with the help of the
new "Visible reviewers" setting in the widget. Available options are
"All", "Team" and "Self". If there are many reviewers involved, this
allows to reduce visual noise. At the same time, for large tables, this
improves performance.