Tracing overlay: suspend updates during most node/link creation actions
With the exception of node insertion and simple node appending, all
node/connector creation and linking operations now suspend the layer so
that it won't update during the respective operation. Afterwards, the
suspension mode is reset to its original state. Node insertion and node
appending are handled separately.

Also update context menu to provide information to hide handler whether
the menu is hidden after a selection or without a selection.

This removes one more node update mentioned in catmaid/CATMAID#1594.