Tracing layer: check selected mirror if auto-switch is turned on
The tile layer has a setting for whether the next accessible mirror
should be selected, if the current one isn't accessible. If this setting
is changed from off to on, a canary test for the selected mirror is now
initiated. This action is probably what most people expect when turning
this setting on.