Pixi layer: prevent caching of invalid textures caused by loading error
If an image can't be loaded, e.g. due to its hosts being inaccessible,
it is marked as invalid. Regardless, Pixi stores invalid textures in its
texture cache. If a tile has been queued to be loaded and is present in
the texture cache, but is marked as invalid, CATMAID assumes it is still
loading. In case of a loading error, this breaks down, causing images
that couldn't get loaded to remain in this "still loading" state. This
in turn causes a synchronized layer update to wait, which effectively
prevents a stack viewer update on sections that have already been
visited and that contain tiles that couldn't be loaded.

This change causes tiles that couldln't be loaded to be removed from
Pixi's cache and to not be respected for reference counting in CATMAID's
texture manager. This makes it possible to browse tracing data even if
images are not accessible.

There are other reasons texture can be invalid, e.g. during loading.
This however shouldn't matter since the place a texture is now tested
for validity is part of the callback executed once a texture in done
loading (successful or not).