Graph widget: accelerate by a lot the Find paths function under the "Grow" tab.
Turns out the cutoff for networkx.all_simple_paths function is the number
of hops and not the number of vertices in the path (one more than the numbe
r of hops).

The claimed 'acceleration' is merely that the cost of finding paths of
length N is now what used to be the cost of finding paths of length N-1.
Given the non-linear nature of the cost of path search, this can be a lot.