3D Viewer: fix memory leak in textured Z plane
With every new section, a new set of textures, images and callback
functions was created before. It seems as if the callback functions
prevented the image element objects to be garbage collected, leading to
a memory leak. This has been fixed by reusing existing Z section
textures and images if they are available. Instead of creating new
callbacks with every section, a single callback is now used and needed
state is appended to the image object (which allows access to it in
callback).

Fixes catmaid/CATMAID#1605