Tracing overlay: fix visibility of new connectors w/ visibility groups
If visibility groups are in use, connectors are displayed only if one of
their connected nodes is visible. This test, however, is only done once
and the result is cached. To reduce the number of updates, connector
links don't enforce a node reload anymore, which prevents recalculation
of this visibility information and newly created connectors are not
shown anymore. To fix this, the cached visibility group information is
reset and recalculated. This is implemented as an optional flag for the
updateVisibility() function on treenodes and connectors, which allows
bypassing the cache.