3D viewer: display all visible tile layers as Z sections
Tile layers are rendered currently simply blended together (in a
separate scene) so that the layer further up is shown. Currently, no
combination with layers below happens. For each layer the respective
mirror from the selected stack viewer is used. The rendering is
implemented in a way that makes adding filters and other blend modes at
a later point in time not very hard. All layers are blended together in
an extra scene and the resulting render is screen space mapped onto the
z section in the main scene (which is only a simple quad now).

Also, the Z plane isn't re-created anymore if no options changed, which
helps update performance. And a GPU memory leak wrt. to z plane textures
has been fixed.