Export widget: neurons can now be exported as NRRD files
The NRRD export is based on @jefferis' NAT R package to allow template
space transformation. To set this up properly, follow the directions on
[1], which essentially require running the following in R either
globally or for the user running CATMAID on the server:

  if (!require("devtools")) install.packages("devtools")
  devtools::install_github("jefferis/rcatmaid")

This also requires CMTK [2] to be installed. It appears easies to just
download the source and compile. The resulting binary folder has to be
made accessible in /usr/local/lib/cmtk/bin and rcatmaid will pick it up.

By default, this export works synchronously, but depending on the size of
a neuron and the available server resources, asynchronous processing
might be needed, which can be selected in the front-end dialog. A Celery
task is provided, which writes a result message after it is done, which
contains a download link.

The back-end creates an R script to do the export. Currently, this
script needs to connect separately to CATMAID and needs additional
settings to do this:

  CATMAID_FULL_URL = ""
  CATMAID_HTTP_AUTH_USER = None
  CATMAID_HTTP_AUTH_PASS = None

Adjust these to match your setup. If no HTTP authentication is used,
leave both respective settings None. The full URL is expected to be the
fully qualified CATMAID URL as it is seen from the outside.

[1] https://github.com/jefferis/rcatmaid
[2] https://www.nitrc.org/projects/cmtk/