Update on_edit() trigger function to also update txid
This migration updates CATMAID's on_edit() trigger function to not only
update the edition time, but also the update's transaction ID. This is
done also by the history tracking update trigger, but history tracking
isn't enabled everywhere. An up-to-date transaction ID allows to relate
table updates to the transaction log table.

The txid column is part of all concept tables, which are the tables on
which the on_edit() trigger is registered.