Tracing tool: add button to refresh annotation and neuron name caches
In a collaborative environment it can easily happen that one user (A)
changes data on which another user's (B) neuron name representation depends.
The added button allows user B to update CATMAID's front-end caches of
annotations and neuron names as well as visibility group information to
the most recent state.