3D viewer: promisify adding of skeletons and queue multiple calls
If append is called multiple times on a 3D viewer, it can make the 3D
viewer's loading act weirdly. The problem is that some asynchronous
parts of both calls can get intermixed. This could be seen when loading
many skeletons into the 3D viewer, hiding all but one, enabling or
disabling smoothing and then changing the color of the visible skeleton
continuously. It is now possible to queue multiple append actions.

This also removes some duplicate colorization actions.