Neuron history: improve performance by avoiding string compare
All events need to be sorted to be analyzed. If both dates to compare
are equal the compare function used to do a string comparison of the
original string timestamps using localeCompare. This is done, because
Postgres supports a higher timestamp precision. This precision however
is not really needed in the neuron history widget (microseconds in
Postgres over milliseconds in JavaScript). Therefore, a simpler
comparison is done now: only the dates are compared, no fallback to
string comparison. This improves performance considerably.