Fix freezing of CATMAID after user actions without server
If the server is unavailable, for instance due to a disconnected
network, the tracing layer was able to freeze the update of all stacks
after a user action like a section change. Neither image data nor
tracing data would change anymore until the tracing tool was reloaded.
The reason for this was a NetworkAccessError that the request queue
throws in its state change handler. This exception can only be caught
through the global error handler and so far a caller couldn't know this
happened. Like it is done for server errors, the callback is now
executed as well when the network becomes disconnected. The Service
Unavailable error (503) is presented to the caller. Ideally, we would
have a dedicated error callback for this (or we would finally use
promises also internally for fetching data.

This, however, is unrelated to catmaid/CATMAID#1619, because there the
image data can still be navigated and only the tracing data froze.