3D viewer: fix active node split in changed virtual nodes
If a skeleton is loaded into the 3D Viewer and a virtual node is
selected, the active split shading will try to calculate the coloring
weights between both real nodes so that the color gradient center is
close to the virtual node. To do this it needs the location of its real
child and parent node. If the virtual node changes its child or parent
in the tracing layer, for instance to a materialized virtual node
between itself and its current real parent, the active node split
shading will try to get the location of the new parent node. Without a
reload, however, this information is not available in the 3D viewer and
the active node split shading would error. This commit makes the active
node split shading handle this situation more gracefully: If no
parent/child information is available, no virtual node weighting is
done.

Thanks to @aes123 for providing a stack trace and additional input on
this!

See catmaid/CATMAID#1619