Add option to toggle checkboxes based on a mouse drawn rectangle
Using "Shift + x" or "Ctrl + x" will now toggle a checkbox selection
mode and the cursor turns into a crosshair. In this mode one can draw a
rectangle everywhere on the screen and all checkboxes enclosed by the
rectangle will be checked (Shift + x) or toggled (Ctrl + x). Either
releasing the drawing mouse click or a second "Shift + x" or "Ctrl + x"
will deactivate the tool again.

This makes it much faster to select for instance a subset of the neurons
visible in one widget or selecting multiple options at once.

Closes catmaid/CATMAID#1515