Node back-end: add two new node providers that test only bounding boxes
The new node providers can be used with the keys 'postgis2dblurry' and
'postgis3dblurry'. Both node providers test only for bounding box
intersection of edges with the field of view. This allows for more false
positives and therefore a node limit likely has to be increased. The
client will of course still only render nodes that are actually in the
field of view. Especially for large datasets this less precise test can
improve performance.