Tracing overlay: make catch radius around Z key node creation smaller
When using the Z key to create new nodes, existing nodes are not as
easily selected from a distance anymore. Before, the radius around a
click event was too large and the closest node around the click was
selected. Other nodes will now only be selected if the mouse cursor is
close to their circle graphic on screen.

Since nodes are available in stack space, the catch radius around the
mouse cursor should be in stack space, too. Before, a project space
catch space lookup around the node position in stack space was done.
This cause the catch radius to become bigger and other nodes were
selected more easily if they were closer than the already selected node.