Project statistics: fix cable length click action
When the cable length traced by a user in the statistics table is
clicked, a selection table with skeletons that were modified is opened.
This was broken due to a missing parameter set in a recent updated of
the corresponding skeleton request. All skeletons were requested, which
ultimately made the review and annotation queries too big.