Settings widget: fix settings widget corner case error
Depending on how Nginx is configured, a missing trailing slash in a URL
was a problem or not. The correct URL for retrieving a list of volumes
has a trailing slash, which is now fixed and makes volume retrieval in
the settings widget work regardless of the Nginx configuration.

This removes also the last direct use of requestQueue from the Settings
Widget.

Fixes catmaid/CATMAID