Tracing overlay: fix freezing due to overlapping actions
Not all tracing actions are currently queued correctly. Creating new
connector nodes and adding partner nodes is an example for this. This
needs to be fixed, but in a separate commit. This only fixes the effect
that two overlapping actions can cause the tracing layer to remain in
suspended mode, which effectively freezes it. When a user interaction
on the tracing layer is handled, the current state of the suspend mode
is stored and reset after all operations are finished. For a new
interaction after another one has been started, this means it stores the
information that layer is suspended (due to the first interaction) and
will reset it after its operation back to that state (after suspending
it as well). If the second action lasts longer than the first one (which
is likely), the layer will remain in a suspended state, because it was
reset to it.

This fix will make sure that the tracing layer is always reset to being
unsuspended after a user interaction and not to the original value. This
should not conflict with any use cases. Additionally, the error handling
for tracing interaction is now better.

Issue catmaid/CATMAID#1619 should be fixed by this, but the correct
queuing of connector related user interaction needs to be fixed still.