Add skeleton summary table and update edge trigger
The new table catmaid_skeleton_summary keeps track of each skeleton's
original creation time, last edition time, cable length and node count.

This table is kept up-to-date by statement level triggers in the
database. This requires a new minimum Postgres version of 10. The
reason for using statement level triggers is performance. For large
skeletons, calling a trigger function for each node is a big overhead.
In contrast, statement level triggers are only executed once per
statement.

A downside of statement level triggers is that they are not executed for
child tables when inheritance is used. Therefore, the node location
update had to be changed to explicitly refer to treenode or connector
updates, in order to update the respective summary and edge tables
correctly. It would be best if we have some mechanism in place that
prevents direct writes to location

The summary update is performed together with the edge table update,
because similar information is needed. Therefore, the current edge
computation trigger is replaced.

Additionally, this implementation fixes a bug of the former edge table
update. In the past only the edge from a changed node to its parent was
update. Now child edges are updated as well, both for the edge table and
the summary.

Performance is pretty much the same as before. This is in fact an
improvement, because more work is now done (summary update, child
edges).

A TRUNCATE trigger on the summary table is not really needed, because in
the rare event that a treenode table is truncated, this likely happens
manually. This now would require to also explicitely truncate
catmaid_skeleton_summary.