Move coveralls config file to searched location
This location changed when Django's manage.py file was moved up one
level. This made coveralls look at migrations and tests as well, which
resulted in a lower coverage. With the config file in its expected
location, the test coverage goes above 50% again and tests complete a
little bit faster.