Reconstruction sampler: fix incomplete connector display in synapse step
Users are expected to change the selected interval in the synapse step
of the reconstruction sampler. Therefore, no cached arbor instance can
be used when displaying connectors of the selected interval. The widget
will now reload the complete arbor to ensure the latest skeleton version
is used to filter connector nodes.