Navigator: update section change handling
The Slider class is updated so that it now supports an optional custom
validation function and a prev/next step transformer. This allows for
smoother major step jumps (because the navigator doesn't have to correct
the Z values that are not valid with the correct settings) and it allows
for easier section change pattern modification. This also fixes a bug
due to which it was possible that other selected stacks are ignored
when going back and multiple broken slice check stacks are selected.