Review: fix Shift + W if first unreviewed node is virtual
If the first unreviewed real node is too far away from the last reviewed
real node, Shift + W will select the first virtual node in the
respective segment. The Review Widget didn't correctly set the last
index of the node coming before the newly selected node if the newly
selected node was a virtual node. This is fixed now.