Graph widget: add support for proportion based edge labels
This makes edge labels configurable through the properties dialog.
Besides the 'absolute connection count' strategy used so far, there are
two new strategies available "Fractions of outbound connections" and
"Fractions of inbound connections". Instead of an absolute number they
display the relative fraction of the number of the number of synapses
represented by a particular edge to the total number of synapses of this
kind (inbound or outbound) on the respective node.