Tracing overlay nodes: fix node movement threshold
In a tracing layer, nodes are stored in stack coordinates. The test
whether a node moved far enough works also based on stack coordinates
(that's the space the dragging distance is calculated in), but uses the
wrong scaling: node.scaling, which multiplies node.stackScaling with the
stack's max(X,Y) resolution. Using the resolution there accidentally
works more or less with resolutions close to one (e.g. 4 nm/px in FAFB),
but breaks down quickly the smaller the resolutions gets (i.e. the more
nm/px). For the test whether a node should be moved, this means that
more and more closer points are rejected, which requires the user to
drag a node very far before it is considered a move. Scaling the moved
distance without resolution fixes this.