Python 3: fix git based version retrieval
In Python 3, the data returned by subprocess.communicate() is no string
unless the text mode is used. Text mode can be enforced by various
means, among them accepting universal newlines. This is done here,
because it works for both Python 2 and 3.

A test for this has already been added in last commit.