Tracing layer: don't set active node null before deleting it
Setting the active node null also causes the active skeleton to become
null. This was a problem for instance in the 3D Viewer, where as part of
handling a deleted node the active skeleton is updated. This happens
however only if the active skeleton is the skeleton of the deleted node
and this can't be checked anymore if the active node was set null prior
these events.