Tracing layer: fix active node sync between (ortho) views
To allow CATMAID to start a node action in one stack viewer and finish
it in another one, the active node is always guaranteed to be available
from all tracing layers (and hence stack viewer). To allow this also
between stack viewers with stacks of different orientation, the active
node coordinates are first transformed into project space and then into
the stack space of the target tracing layer. When converting to project
space, the X coordinate of the node was ignored and the Z component used
instead. This caused wrong node locations for those imported active
nodes. This was especially visible when selecting a virtual node in one
orientation (e.g. ZY) and then split the skeleton at that location. The
virtual node is materialized based on its location and the split
dialog's 3D Viewer showed a completely off node location. This is fixed
now.