Patch GEOS version loading to be compatible with Django 1.10
The GEOS version representation changed and now includes a version hash.
Django 1.10 won't be patched to support this new version. So to be able
to continue using Django 1.10 for now, the version parsing needs to be
monkey patched. The version parsing happens during import time and
therefore we can't just simply monkey patch the module. Instead we
create a custom loader class which intercepts loading the module

  django.contrib.gis.geos.libgeos

and patches the version parsing. After this is done, the custom importer
is removed again.