Tracing layer: instantiate virtual nodes after OK in split dialog
So far, if a split was performed on a virtual node, the virtual node was
instantiated before the split dialog was shown. This would leave one was
an instantiated node even if one canceled the split operation. This
changed and now the virtual node is only created once the user presses
OK in the dialog. Canceling will leave the virtual node alone.