Make default ortho layouts configurable
The new settings definition in CATMAID.Layout.Settings has a single
field: default_layouts. It allows to specify a list of layout
definitions, which are tried to apply on the current window
configuration. Each layout definition is a string that mimics function
invocations for v(a, b), representing a vertical split; h(a, b),
representing a horizontal split; and o(a) which represents an optional
regular window. In such a definition, v() and h() accept either
orthogonal views (XY, XZ, ZY) or regular windows (F1), which can
optionally be optional (o(F1)). The default layout for all three
orthogonal views looks like this:

  h(v(XY, XZ), v(ZY, o(F1)))

This requires all three orthogonal views and will create a F1 (help)
window if no other regular window is already available.

Other definitions can be added and the first matching one will be used.
Being able to define multiple patterns is useful for projects with
multiple stack views of different orientation combinations, e.g. a stack
group with XZ and ZY and another one with all three orientations.