Layouts: apply layouts also to stacks loaded without stack group
This makes it possible to have default layouts for regular single stack
configurations. For instance, to always open a Neuron Search widget
right next to a newly opened stack, this layout spec can be used:

  h(XY, o("neuron-search"))

A layout is not applied if additional stacks are loaded through the
stacks menu. Views opened through a URL will have the first matching
layout applied.