Project initialization: update stack menu only once
So far the stack menu was updated on every stack viewer opening. This
was a problem when opening stack groups, because for each member stack
the list of stacks was queried again and the stack menu updated. Instead
the stack menu is now only updated once, when the project is created.