Add basic WebSocket back-end support
This adds Channels [1] to provide WebSockets functionality. Endpoints
that can be used with WebSockets are all mapped under the /channels/ URL
component and are defined in routing.py for each app. Currently the only
available endpoint is

  /channels/updates/

which will add the connecting user to a user specific group. This group
can the be used by e.g. the back-end to push messages to the client.

This setup works with runserver out of the box, but for a production
setup all requests to a /channels/ URL would need to be proxied to an
ASGI server like Daphne (which is already installed as part of
channels).

As a default back-end, the ASGI IPC [2] back-end is used, because it
allows interprocess communication on the same system and doesn't require
an extra server process. Other servers like Redis or RabbitMQ can be
used as well.

[1] https://channels.readthedocs.io
[2] https://channels.readthedocs.io/en/stable/backends.html#ipc