Add grid based cache option for spatial queries
Similar to the simple node_query_cache table and node providers, there
is now also a node_grid_query_cache table and node provider. Together
with the node_grid_query_cache_cell table, a cache for spatial data in
individual AABB bounding boxes in a uniform grid can be be stored and
queried with this. The management command "catmaid_update_cache" has
been updated sp that it can work with grid caches, too. A typical call
to create new grid cache cells could look like this:

  ./manage.py catmaid_update_cache_tables --project=1 --cache grid \
      --type msgpack --cell-width 50000 --cell-height 50000 --cell-depth 40

This will create a uniform grid with cells of size 50000x50000x40 nm.
Other parameters like --min-z/--max-z or --orientation can be used as
well.

The optional settings parameter DEFAULT_CACHE_GRID_CELL_WIDTH,
DEFAULT_CACHE_GRID_CELL_HEIGHT and DEFAULT_CACHE_GRID_CELL_DEPTH allow
to define defaults for the above management command.

Functionality to update cells automatically is included as well. This
uses Postgres' NOTIFY/LISTEN feature, which allows for asyncronous
events. To lower the impact on regular tracing operations (especially
joining), an insert/update/delete trigger for treenode and connector
will execute a conditional trigger function, which is set on CATMAID
startup. Surprisingly, there doesn't seem to be a big impact of
pg_notify as long as it is done in a separate function. In case these
events ("catmaid.spatial-update" and "catmaid.dirty-cache") should have
an impact, it is possible to disable them by setting
`SPATIAL_UPDATE_NOTIFICATIONS = False` in `settings.py`. It is enabled
by default, because no large differences could be measured.

Cache updates work by running two additional worker process in the form
of management commands: `catmaid_spatial_update_worker` and
`catmaid_cache_update_worker`. The former is responsible for listening
to the "catmaid.spatial-update" Postgres event and adds entries to the
table `dirty_node_grid_cache_cell` for each intersected cache cell in an
enabled grid cache. Upon inserts and updates this table issues the
"catmaid-dirty" cache event, which the second managment command will
listen to. It's its responsibility to update the respective cache
cells and remove entries from the dirty table.

When treenodes are created, moved or deleted the database emits now the
event "catmaid.spatial-update" along with the start and end node
coordinates. The same happens with changed connectors and connector
links. Other processes can use this to asynchronously react to
those events without writing to another table or blocking trigger
processing in other ways. This will be used to to rebuild caches.

The front-end allows to set a "Level of detail" (LOD) value in the
tracing layer settings. By default, this is set to "max", which causes
all LOD levels to be included. Setting this to 1, will include only the
first level.

To use levels of detail, the command line options are available:

--lod-levels to configures over how many levels the data should be
             spread, default 1.
--lod-bucket-size to configure the how many nodes fit into the first LOD
                  level. The size of Additional buckets/levels depends
                  on the selected LOD strategy. Default: 500.
--lod-strategy to configure whether the increase of visible data per LOD
               level should be "linear" or "quadratic". If the former,
               each LOD level will have "lod-bucket-size" more data per
               level. For quadratic, each level has space for
               "lod-bucket-size" to the power of (LOD level) node
               available. Default is "quadratic".

These settings can be used e.g. like this:

  ./manage.py catmaid_update_cache_tables --project=1 --cache grid \
      --type msgpack --cell-width 50000 --cell-height 50000 --cell-depth 40 \
      --lod-levels 50 --lod-bucket-size 5 --lod-strategy quadratic